Fix start-date bug
Fixes a bug where the start date was not respecting the `since`
attribute in the query string parameters because it was always
checking for `window.CHART_START_DATE` first, which is statically
defined in the HTML and always present. This means it was always
ignoring the query string because it always had a value to take.

Also updates the statically defined start date to be Jan. 01 2023